move coverage into oas
